requirements.txt
streamlit==1.38.0
google-generativeai==0.8.3
qdrant-client==1.12.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
sqlite-utils==3.37
scrapy==2.11.1
trafilatura==1.12.2
jaydebeapi==1.2.3
slack-sdk==3.27.1
atlassian-python-api==3.41.3
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain>=0.1.0